Fixed Sidebar References
Apparently html <base> paths ruin the ability to use empty anchor
links. They anchor to the base path, ignoring the relative current
path. I'm not sure if there is a local way to solve this, but for
now i simply combined the current url to all anchor links *(in
the sidebar, at least)*.