Removed Forking Guides
Since i am on the subject of forking, i decided to remove the
Github guide. Why? It's just too big of a subject, and while we may
have had an okay guide on it, i think it would be better to simply
not cover it.

If someone wants to contribute, they should know how to PR. If they
don't, they can send me the article.